requirements.txt
pytorch
torchvision
numpy
tqdm
numba
matplotlib
opencv-python
detectron2